Backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
SQLAlchemy
streamlit
schedule
apscheduler
requests
sendgrid
pandas
matplotlib
seaborn
PyJWT
passlib
bcrypt 
aiosqlite
pytest
pytest-asyncio